Move MapEntries in own folder and add doc
