+++ OsuPlayer/Views/PlaylistView.axaml
@@ -5,6 +5,7 @@
              xmlns:dbReader="clr-namespace:OsuPlayer.IO.DbReader;assembly=OsuPlayer.IO"
              xmlns:valueConverters="clr-namespace:OsuPlayer.ValueConverters"
              xmlns:classes="clr-namespace:OsuPlayer.Data.OsuPlayer.Classes;assembly=OsuPlayer.Data"
+             xmlns:dataModels="clr-namespace:OsuPlayer.IO.DbReader.DataModels;assembly=OsuPlayer.IO"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:Class="OsuPlayer.Views.PlaylistView">
 
@@ -61,7 +62,7 @@
             <ListBox Background="Transparent"
                      Items="{Binding SelectedPlaylist.Songs, Converter={StaticResource PlaylistValueConverter}}">
                 <ListBox.ItemTemplate>
-                    <DataTemplate DataType="dbReader:MinimalMapEntry">
+                    <DataTemplate DataType="dataModels:MinimalMapEntry">
                         <Grid ColumnDefinitions="*, Auto" RowDefinitions="Auto, Auto">
                             <TextBlock Grid.Row="0" Grid.Column="0" Text="{Binding TitleString}" FontSize="20"
                                        FontWeight="SemiBold" />
